In the class EdbScanProc, the method:   EdbScanProc::ReadPatTXT(EdbPattern &pred, int id[4], const char *suffix, int flag0) is splitted in two:   EdbScanProc::ReadPatTXT(EdbPattern &pred, int id[4], const char *suffix, int flag0) and:   EdbScanProc::ReadPatTXT(const char *file, EdbPattern &pred, int flag0) the latter to have a direct access to a file
The method is also changed. In the previous version, if the filename contains the 
word "man", it expects 6 columns, otherwise 10 columns (the 4 errors sx,sy,stx,sty).

Now, the decision does not depend to the filename anymore, but to the number of 
columns of the first row. This guarantee the backward compatibility and offers more 
flexibility.



git-svn-id: http://emulsion.na.infn.it/svn/fedra/trunk@778 96872706-0045-0410-9534-84be8447c86c