Three modifications:
I)   Formatted the code using the __astyle__ program. This gives a normed code output using tabs and spaces and bracket according to the official C coding standard.

II)  In EdbDataSet: After Creating EdbPVRec object as sub-object of the EdbDataProcObject it is also given over as private member object, until now it was just deleted afterwards.

III) Added the -wPVR option in recset. This makes it possible to read and write PVR objects of the commonly used data structure to write the Volume into a root file. I do use it for MC-Event batch processing.

---------------------------

ONE IMPORTANT COMMENT AT Valeri.

I tried to compile latest fedra on my notebook using Ubuntu 12.10 and the built in gcc compiler in its standard version 4.7.
This did not work correctly for linking the "app" programs. Switching to gcc-4.4.7 helped solving this problem.

Valeri, could you please check it and see, if you can reproduce this problem? If so, some build-makefiles should be modified. Thanks a lot.

-----------------------------

Best greetings to all fedra users and developers! :-)




git-svn-id: http://emulsion.na.infn.it/svn/fedra/trunk@1312 96872706-0045-0410-9534-84be8447c86c