Added Function EdbSegmentsBox::AddSegmentNoDuplicate( EdbSegP &s )
This function will add the segment s to the current list of
segments checking for duplicates. If any segment in the list is found,
that is compatible to this, the segment will _not_ be added,
instead a warning message will be printed (according to 
verbosity level)

This function might be slow.




git-svn-id: http://emulsion.na.infn.it/svn/fedra/trunk@1410 96872706-0045-0410-9534-84be8447c86c