Important patch in the reconstruction algorithm:
EdbShowAlg_OI::FindPrecedingBTs()
The condition check "if (dZ>(3*1300.0)+30.0)" was supposed to look for a dZ greater 3 plates.
Since experimentally often observed, dZ is about 1350 ... so the upper cut really does not
look for 3 plates, but only for two. Adaption to cut "if (dZ>(3.5*(1300.0+30.0)))" does give
the right number of plates to be looked at.



git-svn-id: http://emulsion.na.infn.it/svn/fedra/trunk@1431 96872706-0045-0410-9534-84be8447c86c