Added the "ShowRec" program to the repository.
This was originally the first program the shower reconstruction
was done. Its main purpose was to compare different
shower reconstruction algorithms and different parameter
sets for the algorithms to find optimum values.
Later on, the best algorithms were implemented in the fedra
shower library.
It can only reconstruct showers. Energy estimation and 
particle identification cannot be done with this program. 

This program is merely added for completeness of the shower code.




git-svn-id: http://emulsion.na.infn.it/svn/fedra/trunk@1475 96872706-0045-0410-9534-84be8447c86c